third_party/WebKit/LayoutTests/fast/events/touch/gesture/touch-gesture-fully-scrolled-iframe-propagates.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../resources/js-test.js"></script>
<script src="resources/gesture-helpers.js"></script>
<style type="text/css">

::-webkit-scrollbar {
    width: 0px;
    height: 0px;
}

#bluebox {
  width: 100px;
  height: 100px;
  background: blue;
  padding: 0px;
  margin: 0px;
}

#outerdiv {
    width: 200px;
    height: 200px;
    overflow-y: scroll;
    overflow-x: scroll;
}

td {
  padding: 0px;
}

</style>
</head>
<body style="margin:0" >

<div id="outerdiv">
  <table border="0" cellspacing="0px" >
    <tr><td>
        <iframe style="display: block;" frameBorder="0" id="touchtargetiframe" src="resources/scroll-inside-editable-iframe.html"></iframe>
    </td></tr>
    <tr><td>
      <div id="bluebox"></div>
    </td></tr>
  </table>
</div>

<p id="description"></p>
<div id="console"></div>
<script type="text/javascript">

var movedbox;
var touchtarget;
var iframe;
var iframeScrollingElement;
var iframeDocumentElement;
var expectedGesturesTotal = 1;
var gesturesOccurred = 0;
var scrollAmountX = ['0'];
var scrollAmountY = ['50'];
var wheelEventsOccurred = 0;
var expectedWheelEventsOccurred = ['0'];
var scrollEventsOccurred = 0;
var scrolledElement = 'movedbox';
var scrollEventsOccurred = 0;
var expectedScrollEventsOccurred = '1';

function firstGestureScroll()
{
    iframe = touchtarget;
    iframeScrollingElement = iframe.contentDocument.scrollingElement;
    
    var amountToScroll = iframeScrollingElement.scrollHeight - iframe.clientHeight;

    iframeScrollingElement.scrollTop = amountToScroll;

    // Sanity - make sure the iframe is actually fully scrolled
    shouldBe('iframeScrollingElement.scrollHeight - iframeScrollingElement.scrollTop', 'iframe.clientHeight');

    eventSender.gestureScrollBegin(10, 72);
    eventSender.gestureScrollUpdate(0, -20);  
    eventSender.gestureScrollUpdate(0, -18);
    eventSender.gestureScrollUpdate(0, -15);
    eventSender.gestureScrollUpdate(0, -10);
    eventSender.gestureScrollEnd(0, 0);

    amountToScroll = movedbox.scrollHeight - movedbox.clientHeight;
    scrollAmountY[0] = amountToScroll.toString();

    // Wait for layout.
    checkScrollOffset();
}

if (window.testRunner)
    testRunner.waitUntilDone();

function runTest()
{
    movedbox = document.getElementById("outerdiv");
    touchtarget = document.getElementById("touchtargetiframe");

    touchtarget.contentDocument.addEventListener("scroll", recordScroll);
    touchtarget.contentDocument.documentElement.addEventListener("mousewheel", recordWheel);

    if (window.eventSender) {
        description('This tests that a fling gesture sent to an iframe with no remaining scroll offset ' +
                    'is correctly targeting the parent container.');
        
        if (checkTestDependencies() && eventSender.gestureScrollUpdate)
            firstGestureScroll();
        else
            exitIfNecessary();
    } else {
        debug("This test requires DumpRenderTree.  Gesture-scroll the page to validate the implementation.");
    }
}
</script>



</body>
</html>
