third_party/WebKit/LayoutTests/fast/events/touch/gesture/touch-gesture-fully-scrolled-iframe-propagates-expected.txt


This tests that a fling gesture sent to an iframe with no remaining scroll offset is correctly targeting the parent container.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS iframeScrollingElement.scrollHeight - iframeScrollingElement.scrollTop is iframe.clientHeight
PASS movedbox.scrollTop is 50
PASS movedbox.scrollLeft is 0
PASS wheelEventsOccurred is 0
scroll event 0+> [object HTMLDocument]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
